// hw1/src/enums/operator-type.ts
/** Тип оператора */
export enum OperatorType {
    /** Вычисляет значения чисел находящихся слева и справа от оператора */
    ComputeSides = 1,
    /** Вычисляет значение числа находящегося справа от оператора*/
    ComputeValueRight = 2,
    /** Вычисляет значение числа находящегося справа от оператора*/
    ComputeValueLeft = 3,
}